Also parse URLs in the Pecl factory
